Add resque implementation
This adds a resque implemenation of an Enqueuer,
and wires up the main worker process
to handle GoReviewJobs.

In theory, this should be sufficient
to apply golint to a Go pull request,
assuming Hound created the necessary GoReviewJob.
