+++ main.go
@@ -6,6 +6,13 @@ import (
 )
 
 func main() {
+	var (
+		reviewer = reviewerFunc(lintReviewer)
+		queue    = newResqueEnqueuer("high")
+	)
+
+	goworker.Register("GoReviewJob", newGoReviewJob(reviewer, queue))
+
 	if err := goworker.Work(); err != nil {
 		log.Error(err)
 	}
